Refine PCI Segment Library to avoid possible intrinsic function.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@7752 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdePkg/Library/PeiPciSegmentLibPciCfg2/PciSegmentLib.c
@@ -42,7 +42,7 @@
 
 **/
 #define PCI_TO_PCICFG2_ADDRESS(A) \
-  ((((A) << 4) & 0xff000000) | (((A) >> 4) & 0x00000700) | (((A) << 1) & 0x001f0000) | (LShiftU64((A) & 0xfff, 32)))
+  ((((UINT32)(A) << 4) & 0xff000000) | (((UINT32)(A) >> 4) & 0x00000700) | (((UINT32)(A) << 1) & 0x001f0000) | (LShiftU64((A) & 0xfff, 32)))
 
 /**
   Gets PCI CFG2 PPI.
